light_kb/requirements.txt
openai
chromadb
sentence-transformers
semantic-chunkers
pymupdf
python-docx
tqdm
numpy
fitz